Update: task db schema

+++ twitkit-fridge/twitkit-fridge-src/src/main/java/com/enkanrec/twitkitFridge/api/form/TidForm.java
@@ -23,5 +23,5 @@ public class TidForm extends JsonDataFridgeForm {
      * 任务id
      */
     @NotNull
-    private Long tid;
+    private Integer tid;
 }
